requirements.txt
discord==1.7.3
webcolors
python-dateutil
tabulate